+++ src/main/java/roomescape/controller/ReservationsController.java
@@ -2,8 +2,8 @@ package roomescape.controller;
 
 import org.springframework.web.bind.annotation.*;
 import roomescape.domain.Reservation;
-import roomescape.domain.ReservationRequestDto;
-import roomescape.domain.ReservationResponseDto;
+import roomescape.dto.ReservationRequestDto;
+import roomescape.dto.ReservationResponseDto;
 
 import java.util.ArrayList;
 import java.util.List;
