+++ src/main/java/roomescape/dto/ReservationRequestDto.java
@@ -1,4 +1,6 @@
-package roomescape.domain;
+package roomescape.dto;
+
+import roomescape.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
